[Fix] Get rid of stupid warning introduced in previous commit.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5875 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/ModInstrument.cpp
@@ -309,7 +309,7 @@ void ModInstrument::Sanitize(MODTYPE modType)
 	LimitMax(nPanSwing, 64u);
 	LimitMax(nVolSwing, 100u);
 
-	Limit(nPPS, -32, 32);
+	Limit(nPPS, int8(-32), int8(32));
 
 	LimitMax(nCutSwing, 64u);
 	LimitMax(nResSwing, 64u);
